--- src/main/java/com/sparta/bapzip/ai/application/GeminiInterface.java
@@ -1,18 +1,0 @@
-package com.sparta.bapzip.ai.application;
-
-import com.sparta.bapzip.ai.application.dto.GeminiRequest;
-import com.sparta.bapzip.ai.application.dto.GeminiResponse;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.service.annotation.HttpExchange;
-import org.springframework.web.service.annotation.PostExchange;
-
-@HttpExchange("/v1beta/models/")
-public interface GeminiInterface {
-
-    @PostExchange("{model}:generateContent")
-    GeminiResponse getCompletion(
-            @PathVariable String model,
-            @RequestBody GeminiRequest request
-    );
-}
