+++ src/main/java/com/sparta/bapzip/ai/application/GeminiService.java
@@ -1,36 +1,29 @@
 package com.sparta.bapzip.ai.application;
 
-import com.sparta.bapzip.ai.application.dto.GeminiRequest;
-import com.sparta.bapzip.ai.application.dto.GeminiResponse;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.google.genai.Client;
+import com.google.genai.types.GenerateContentResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class GeminiService {
 
     public static final String GEMINI_MODEL = "gemini-2.5-flash";
+    public static final String CHARACTER_LIMIT = "\n 답변을 최대한 간결하게 50자 이하로";
 
-    private final GeminiInterface geminiInterface;
+    @Value("${GOOGLEAI_API_KEY}")
+    private String apiKey;
 
-    @Autowired
-    public GeminiService(GeminiInterface geminiInterface) {
-        this.geminiInterface = geminiInterface;
-    }
+    public String getResponse(String prompt){
+        Client client = Client.builder().apiKey(apiKey).build();
 
-    private GeminiResponse getCompletion(GeminiRequest request) {
-        return geminiInterface.getCompletion(GEMINI_MODEL, request);
-    }
-
-    public String getCompletion(String text) {
-        GeminiRequest geminiRequest = new GeminiRequest(text);
-        GeminiResponse response = getCompletion(geminiRequest);
-
-        return response.getCandidates()
-                .stream()
-                .findFirst().flatMap(candidate -> candidate.getContent().getParts()
-                        .stream()
-                        .findFirst()
-                        .map(GeminiResponse.TextPart::getText))
-                .orElse(null);
+        GenerateContentResponse response =
+                client.models.generateContent(
+                        GEMINI_MODEL,
+                        prompt+CHARACTER_LIMIT,
+                        null);
+        return response.text();
     }
 }
