--- src/main/java/com/sparta/bapzip/ai/config/AiConfig.java
@@ -1,32 +1,0 @@
-package com.sparta.bapzip.ai.config;
-
-import com.sparta.bapzip.ai.application.GeminiInterface;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestClient;
-import org.springframework.web.client.support.RestClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-
-@Configuration
-public class AiConfig {
-
-    @Bean
-    public RestClient geminiRestClient(@Value("${GEMINI_BASEURL}") String baseUrl,
-                                       @Value("${GOOGLEAI_API_KEY}") String apiKey) {
-        return RestClient.builder()
-                .baseUrl(baseUrl)
-                .defaultHeader("x-goog-api-key", apiKey)
-                .defaultHeader("Content-Type", "application/json")
-                .defaultHeader("Accept", "application/json")
-                .build();
-    }
-
-    @Bean
-    public GeminiInterface geminiInterface(@Qualifier("geminiRestClient") RestClient client) {
-        RestClientAdapter adapter = RestClientAdapter.create(client);
-        HttpServiceProxyFactory factory = HttpServiceProxyFactory.builderFor(adapter).build();
-        return factory.createClient(GeminiInterface.class);
-    }
-}
