--- src/test/java/com/sparta/bapzip/ai/application/GeminiServiceTest.java
@@ -1,19 +1,0 @@
-package com.sparta.bapzip.ai.application;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GeminiServiceTest {
-
-    @Autowired
-    private GeminiService service;
-
-    @Test
-    void getCompletion() {
-        String text = service.getCompletion("서울 맛집을 추천해줘");
-        System.out.println(text);
-    }
-
-}
